Extract out MapVCDelegate, and it's default implimentation
